Move HeaderOracle into oracle module
