+++ trin-core/src/types/mod.rs
@@ -1,5 +1,6 @@
 pub mod accumulator;
 pub mod block_body;
 pub mod header;
+pub mod oracle;
 pub mod receipts;
 pub mod validation;
