+++ trin-history/src/network.rs
@@ -14,7 +14,7 @@ use trin_core::{
             messages::{PortalnetConfig, ProtocolId},
         },
     },
-    types::validation::HeaderOracle,
+    types::oracle::HeaderOracle,
     utp::stream::UtpListenerRequest,
 };
