+++ trin-state/src/lib.rs
@@ -14,7 +14,7 @@ use trin_core::{
     portalnet::{
         discovery::Discovery, storage::PortalStorageConfig, types::messages::PortalnetConfig,
     },
-    types::validation::HeaderOracle,
+    types::oracle::HeaderOracle,
     utp::stream::{UtpListenerEvent, UtpListenerRequest},
 };
